Add port of Sam Westrick's seam carver.
It's a mostly straightforward port, with the main change being due to
the incompatible definition of ForkJoin.alloc.

Performance is not great because some of the parallelism has too small
a grain size, and it crashes if you try to generate multiple GIF
frames in parallel.

I am confident that it can be made much faster, and I'll work some
more on it.

I am not particularly happy with the large amount of utility libraries
I had to copy.  It would be nice to factor some of this into separate
packages.